A successful first hls generation of fermilab's model.
Managed to get the HLSTestNotebook.ipynb to run some kind of hls
synthesis with Vitis.
There are things missing from this synthesis and more steps, so
should follow some kind of tutorial to complete them.
This was run for fermilab's model because currently our models
have more than 2 inputs, which make them incompatible with hls4ml.